backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
sqlalchemy==2.0.35
aiosqlite==0.20.0
asyncpg==0.30.0
faster-whisper
google-generativeai==0.8.3
azure-cognitiveservices-speech==1.40.0
python-dotenv==1.0.1
httpx==0.27.2
pydantic==2.9.0
pydantic-settings==2.5.0
librosa==0.10.2
soundfile==0.12.1
PyJWT==2.8.0
bcrypt==4.2.0
